Changed Book Fragment layout

+++ app/src/main/java/com/moonlightbutterfly/bookid/repository/database/entities/Rating.kt
@@ -5,6 +5,6 @@ import androidx.room.Entity
 
 @Entity
 data class Rating (
-    @ColumnInfo(name = "ratingValue") val ratingValue: Double? = null,
+    @ColumnInfo(name = "ratingValue") val ratingValue: Float? = null,
     @ColumnInfo(name = "ratingsCount") val ratingsCount: Int? = null
 )
